src/backend/python/requirements.txt
anthropic>=0.7.0
websockets>=11.0
pydantic>=2.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
python-dotenv>=1.0.0
mcp>=1.0.0